@@ -1,6 +1,8 @@
 package callbacks
 
 import (
+	"errors"
+
 	"github.com/bwmarrin/discordgo"
 )
 
@@ -18,6 +20,7 @@ func (handler *Handler) ChannelDelete(session *discordgo.Session, channel *disco
 	guild, err := session.State.Guild(channel.GuildID)
 	if err != nil {
 		handler.Log.WithError(err).Error(channelDeleteEventError)
+
 		return
 	}
 
@@ -29,12 +32,14 @@ func (handler *Handler) ChannelDelete(session *discordgo.Session, channel *disco
 		err = session.GuildRoleDelete(channel.GuildID, role.ID)
 		if err != nil {
 			handler.Log.WithError(err).Error(channelDeleteEventError)
+
 			return
 		}
 
 		err = session.State.RoleRemove(channel.GuildID, role.ID)
-		if err != nil && err != discordgo.ErrStateNotFound {
+		if err != nil && !errors.Is(err, discordgo.ErrStateNotFound) {
 			handler.Log.WithError(err).Error(channelDeleteEventError)
+
 			return
 		}
 
