@@ -3,12 +3,15 @@ package mock_test
 import (
 	"testing"
 
+	"github.com/bwmarrin/discordgo"
 	"github.com/ewohltman/discordgo-mock/mockconstants"
 
 	"github.com/ewohltman/ephemeral-roles/internal/pkg/mock"
 )
 
 func TestNewSession(t *testing.T) {
+	t.Parallel()
+
 	session, err := mock.NewSession()
 	if err != nil {
 		t.Fatal(err)
@@ -39,7 +42,10 @@ func TestNewSession(t *testing.T) {
 		t.Error(err)
 	}
 
-	_, err = session.GuildRoleCreate(mockconstants.TestGuild)
+	_, err = session.GuildRoleCreate(
+		mockconstants.TestGuild,
+		&discordgo.RoleParams{Name: mockconstants.TestRole},
+	)
 	if err != nil {
 		t.Error(err)
 	}
